# Data fixer performance - easy bits