--- a/gradle.properties
+++ b/gradle.properties
@@ -1,1 +1,1 @@
-majorMinor: 5.1
+majorMinor: 6.0