--- a/src/main/java/com/mojang/datafixers/OpticParts.java
+++ b/src/main/java/com/mojang/datafixers/OpticParts.java
@@ -11,20 +11,5 @@ import java.util.Set;
 /**
  * apply(i).sType == family.apply(i)
  */
-public final class OpticParts<A, B> {
-    private final Set<TypeToken<? extends K1>> bounds;
-    private final Optic<?, ?, ?, A, B> optic;
-
-    public OpticParts(final Set<TypeToken<? extends K1>> bounds, final Optic<?, ?, ?, A, B> optic) {
-        this.bounds = bounds;
-        this.optic = optic;
-    }
-
-    public Set<TypeToken<? extends K1>> bounds() {
-        return bounds;
-    }
-
-    public Optic<?, ?, ?, A, B> optic() {
-        return optic;
-    }
+public record OpticParts<A, B>(Set<TypeToken<? extends K1>> bounds, Optic<?, ?, ?, A, B> optic) {
 }