--- a/src/main/java/com/mojang/datafixers/functions/In.java
+++ b/src/main/java/com/mojang/datafixers/functions/In.java
@@ -30,7 +30,7 @@ final class In<A> extends PointFree<Function<A, A>> {
 
     @Override
     public int hashCode() {
-        return Objects.hash(type);
+        return type.hashCode();
     }
 
     @Override