--- a/src/main/java/com/mojang/datafixers/functions/Out.java
+++ b/src/main/java/com/mojang/datafixers/functions/Out.java
@@ -30,7 +30,7 @@ final class Out<A> extends PointFree<Function<A, A>> {
 
     @Override
     public int hashCode() {
-        return Objects.hash(type);
+        return type.hashCode();
     }
 
     @Override