… src/main/java/com/mojang/datafixers/functions/PointFreeRule.java
… src/main/java/com/mojang/datafixers/functions/ProfunctorTransformer.java
… src/main/java/com/mojang/datafixers/optics/IdAdapter.java
… src/main/java/com/mojang/datafixers/optics/Inj1.java
… src/main/java/com/mojang/datafixers/optics/Inj2.java
… src/main/java/com/mojang/datafixers/optics/InjTagged.java
… src/main/java/com/mojang/datafixers/optics/ListTraversal.java
… src/main/java/com/mojang/datafixers/optics/Optic.java
… src/main/java/com/mojang/datafixers/optics/Optics.java
… src/main/java/com/mojang/datafixers/optics/Proj1.java
… src/main/java/com/mojang/datafixers/optics/Proj2.java
… src/main/java/com/mojang/datafixers/types/Func.java
… src/main/java/com/mojang/datafixers/types/Type.java
… src/main/java/com/mojang/datafixers/types/families/RecursiveTypeFamily.java
… src/main/java/com/mojang/datafixers/types/templates/Check.java
… src/main/java/com/mojang/datafixers/types/templates/CompoundList.java
… src/main/java/com/mojang/datafixers/types/templates/Const.java
… src/main/java/com/mojang/datafixers/types/templates/Hook.java
… src/main/java/com/mojang/datafixers/types/templates/List.java
… src/main/java/com/mojang/datafixers/types/templates/Named.java
… src/main/java/com/mojang/datafixers/types/templates/Product.java
… src/main/java/com/mojang/datafixers/types/templates/RecursivePoint.java
… src/main/java/com/mojang/datafixers/types/templates/Sum.java
… src/main/java/com/mojang/datafixers/types/templates/Tag.java
… src/main/java/com/mojang/datafixers/types/templates/TaggedChoice.java
… src/main/java/com/mojang/datafixers/util/Either.java
… src/main/java/com/mojang/serialization/Codec.java
… src/main/java/com/mojang/serialization/DataResult.java
… src/main/java/com/mojang/serialization/Decoder.java
… src/main/java/com/mojang/serialization/Dynamic.java