@Gegy 2023-01-17
@bagucode Thanks for looking through it! 🙂 We have run a basic consistency test on this PR and the follow-up by dumping inputs and outputs when upgrading an old world, and verifying against that with the new changes (discovered in the process one datafixer that generates random values 🥲)