@Gegy 2023-02-02
Force-push: dropped change from optional -> nullable after feedback (the overhead is now negligible given changes in #72). Also dropped the `lastStableRule` cache change, as that is no longer used as of #72.